refactor(stores): adopt SerializableModel<T> for type-safe JSON deserialization
- Replace FeedPost::fromJson() pattern with SerializableModel<T>::createFromJson()
- Provides type-safe Outcome<T> results instead of raw parsing
- Proper error handling with Result type
- Apply to Feed.cpp, Search.cpp, EntityStore.h, AggregatedFeedGroup
- Apply to UI components (PostsFeed, SavedPosts, ArchivedPosts, MessageThread)
- Update FeedDataTest with new deserialization pattern
- Improves error handling and validation across data layer

This refactoring sets foundation for:
- Action builders (type-safe state changes)
- Effect layer (explicit side effects handling)
- Better testability of data transformations

+++ plugin/src/ui/animations/AnimationController.cpp
@@ -455,8 +455,8 @@ std::shared_ptr<TransitionAnimation<float>> AnimationController::createScaleAnim
   auto originalBounds = component->getBounds();
 
   animation->onProgress([component, originalBounds, originalWidth, originalHeight](float scale) {
-    int newWidth = static_cast<int>(originalWidth * scale);
-    int newHeight = static_cast<int>(originalHeight * scale);
+    int newWidth = static_cast<int>(static_cast<float>(originalWidth) * scale);
+    int newHeight = static_cast<int>(static_cast<float>(originalHeight) * scale);
     int newX = originalBounds.getX() + (originalWidth - newWidth) / 2;
     int newY = originalBounds.getY() + (originalHeight - newHeight) / 2;
